PWN-5019 - Moved mapping in repository
