+++ app/src/main/java/org/p2p/wallet/solend/interactor/SolendDepositsInteractor.kt
@@ -1,24 +1,12 @@
 package org.p2p.wallet.solend.interactor
 
-import org.p2p.wallet.solend.repository.SolendDepositsRepository
 import org.p2p.wallet.solend.model.SolendDepositToken
-import org.p2p.wallet.solend.model.SolendDomainMapper
-import org.p2p.wallet.user.repository.UserLocalRepository
+import org.p2p.wallet.solend.repository.SolendDepositsRepository
 
 class SolendDepositsInteractor(
-    private val repository: SolendDepositsRepository,
-    private val userLocalRepository: UserLocalRepository,
-    private val mapper: SolendDomainMapper
+    private val repository: SolendDepositsRepository
 ) {
 
-    suspend fun getAllDeposits(tokenSymbols: List<String>): List<SolendDepositToken> {
-        val marketsInfo = repository.getSolendMarketInfo(tokenSymbols)
-        val activeDeposits = repository.getDeposits()
-
-        return marketsInfo.mapNotNull { info ->
-            val tokenData = userLocalRepository.findTokenDataBySymbol(info.tokenSymbol) ?: return@mapNotNull null
-            val activeDeposit = activeDeposits.find { it.depositTokenSymbol == info.tokenSymbol }
-            mapper.toDepositToken(tokenData, info, activeDeposit)
-        }
-    }
+    suspend fun getUserDeposits(tokenSymbols: List<String>): List<SolendDepositToken> =
+        repository.getUserDeposits(tokenSymbols)
 }
