+++ app/src/main/java/org/p2p/wallet/solend/model/SolendDepositMapper.kt
@@ -2,8 +2,30 @@ package org.p2p.wallet.solend.model
 
 import org.p2p.wallet.sdk.facade.model.SolendMarketInformationResponse
 import org.p2p.wallet.sdk.facade.model.SolendUserDepositResponse
+import org.p2p.wallet.user.model.TokenData
 
 class SolendDepositMapper {
+
+    fun fromNetwork(
+        tokenData: TokenData,
+        marketInfo: SolendMarketInfo,
+        activeDeposit: SolendUserDeposit?
+    ): SolendDepositToken = if (activeDeposit != null) {
+        SolendDepositToken.Active(
+            tokenName = tokenData.name,
+            tokenSymbol = tokenData.symbol,
+            iconUrl = tokenData.iconUrl,
+            supplyInterest = marketInfo.supplyInterest,
+            depositAmount = activeDeposit.depositedAmount
+        )
+    } else {
+        SolendDepositToken.Inactive(
+            tokenName = tokenData.name,
+            tokenSymbol = tokenData.symbol,
+            iconUrl = tokenData.iconUrl,
+            supplyInterest = marketInfo.supplyInterest
+        )
+    }
 
     fun fromNetwork(response: SolendUserDepositResponse): SolendUserDeposit =
         SolendUserDeposit(
