+++ app/src/main/java/org/p2p/wallet/solend/repository/SolendDepositsRemoteRepository.kt
@@ -4,8 +4,9 @@ import org.p2p.wallet.infrastructure.network.provider.TokenKeyProvider
 import org.p2p.wallet.sdk.facade.SolendSdkFacade
 import org.p2p.wallet.sdk.facade.model.SolendPool
 import org.p2p.wallet.solend.model.SolendDepositMapper
+import org.p2p.wallet.solend.model.SolendDepositToken
 import org.p2p.wallet.solend.model.SolendMarketInfo
-import org.p2p.wallet.solend.model.SolendUserDeposit
+import org.p2p.wallet.user.repository.UserLocalRepository
 import org.p2p.wallet.utils.Base58String
 import org.p2p.wallet.utils.toBase58Instance
 import timber.log.Timber
@@ -13,7 +14,8 @@ import timber.log.Timber
 class SolendDepositsRemoteRepository(
     private val solanaFacade: SolendSdkFacade,
     private val ownerAddressProvider: TokenKeyProvider,
-    private val mapper: SolendDepositMapper
+    private val mapper: SolendDepositMapper,
+    private val userLocalRepository: UserLocalRepository
 ) : SolendDepositsRepository {
 
     private val currentSolendPool = SolendPool.MAIN
@@ -21,15 +23,25 @@ class SolendDepositsRemoteRepository(
     private val ownerAddress: Base58String
         get() = ownerAddressProvider.publicKey.toBase58Instance()
 
-    override suspend fun getDeposits(): List<SolendUserDeposit> = try {
-        val response = solanaFacade.getAllSolendUserDeposits(ownerAddress, currentSolendPool)
-        response.map { mapper.fromNetwork(it) }
-    } catch (e: Throwable) {
-        Timber.e(e, "Error loading user deposits")
-        emptyList()
+    override suspend fun getUserDeposits(tokenSymbols: List<String>): List<SolendDepositToken> {
+        val marketsInfo = getSolendMarketInfo(tokenSymbols)
+
+        val deposits = try {
+            val response = solanaFacade.getAllSolendUserDeposits(ownerAddress, currentSolendPool)
+            response.map { mapper.fromNetwork(it) }
+        } catch (e: Throwable) {
+            Timber.w(e, "Error loading user deposits")
+            emptyList()
+        }
+
+        return marketsInfo.mapNotNull { info ->
+            val tokenData = userLocalRepository.findTokenDataBySymbol(info.tokenSymbol) ?: return@mapNotNull null
+            val activeDeposit = deposits.find { it.depositTokenSymbol == info.tokenSymbol }
+            mapper.fromNetwork(tokenData, info, activeDeposit)
+        }
     }
 
-    override suspend fun getSolendMarketInfo(tokens: List<String>): List<SolendMarketInfo> {
+    private suspend fun getSolendMarketInfo(tokens: List<String>): List<SolendMarketInfo> {
         val response = solanaFacade.getSolendMarketInfo(tokens, currentSolendPool.poolName)
         return mapper.fromNetwork(response)
     }
