+++ app/src/main/res/layout/fragment_solend_earn.xml
@@ -68,12 +68,13 @@
     </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
 
     <FrameLayout
-        android:id="@+id/shimmerView"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintTop_toBottomOf="@id/textViewTokensTitle">
 
-        <include layout="@layout/view_history_shimmer" />
+        <include
+            android:id="@+id/shimmerView"
+            layout="@layout/view_history_shimmer" />
     </FrameLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
